Add recitation tracks listing API and refactor files for better separation of concerns
