+++ apps/content/api/portal/recitation_tracks.py
@@ -1,0 +1,87 @@
+from typing import Literal
+
+from ninja import Schema
+from ninja.pagination import paginate
+from rest_framework.exceptions import PermissionDenied
+from rest_framework.generics import get_object_or_404
+
+from apps.content.models import Asset, RecitationSurahTrack
+from apps.content.repositories.recitation_track import RecitationTrackRepository
+from apps.core.ninja_utils.auth import ninja_jwt_auth
+from apps.core.ninja_utils.errors import ItqanError, NinjaErrorResponse
+from apps.core.ninja_utils.request import Request
+from apps.core.ninja_utils.router import ItqanRouter
+from apps.core.ninja_utils.tags import NinjaTag
+
+router = ItqanRouter(tags=[NinjaTag.RECITATIONS])
+
+
+class RecitationTrackOut(Schema):
+    id: int
+    surah_number: int
+    audio_url: str | None
+    duration_ms: int
+    size_bytes: int
+    filename: str | None
+
+    @staticmethod
+    def resolve_audio_url(obj: RecitationSurahTrack) -> str | None:
+        if obj.audio_file:
+            return obj.audio_file.url
+        return None
+
+    @staticmethod
+    def resolve_filename(obj: RecitationSurahTrack) -> str | None:
+        return obj.original_filename
+
+
+@router.get(
+    "assets/{asset_id}/recitation-tracks/",
+    response={
+        200: list[RecitationTrackOut],
+        401: NinjaErrorResponse[Literal["authentication_error"], Literal[None]],
+        403: NinjaErrorResponse[Literal["permission_denied"], Literal[None]],
+        404: NinjaErrorResponse[Literal["asset_not_found"], Literal[None]],
+    },
+    auth=ninja_jwt_auth,
+)
+@paginate
+def list_tracks(request: Request, asset_id: int):
+    if not request.user.is_staff:
+        raise PermissionDenied("Staff only")
+
+    asset = get_object_or_404(Asset, id=asset_id)
+
+    repo = RecitationTrackRepository()
+    return repo.get_recitation_tracks_by_asset_id(asset_id=asset.id)
+
+
+class DeleteTracksIn(Schema):
+    track_ids: list[int]
+
+
+@router.delete(
+    "recitation-tracks/",
+    response={
+        204: None,
+        401: NinjaErrorResponse[Literal["authentication_error"], Literal[None]],
+        403: NinjaErrorResponse[Literal["permission_denied"], Literal[None]],
+    },
+    auth=ninja_jwt_auth,
+)
+def delete_tracks(request: Request, data: DeleteTracksIn):
+    if not request.user.is_staff:
+        raise PermissionDenied("Staff only")
+
+    repo = RecitationTrackRepository()
+    tracks = repo.get_recitation_tracks_by_ids(data.track_ids)
+
+    if not data.track_ids or len(tracks) != len(data.track_ids):
+        raise ItqanError(
+            error_name="track_not_found",
+            message="Some track IDs are invalid or do not exist.",
+            status_code=400,
+        )
+
+    repo.delete_recitation_tracks(tracks)
+    return 204, None
