+++ apps/content/api/portal/recitation_tracks_upload.py
@@ -7,7 +7,6 @@ from rest_framework.exceptions import PermissionDenied
 from rest_framework.generics import get_object_or_404
 
 from apps.content.models import Asset
-from apps.content.repositories.recitation_track import RecitationTrackRepository
 from apps.content.services.admin.asset_recitation_audio_tracks_direct_upload_service import (
     AssetRecitationAudioTracksDirectUploadService,
 )
@@ -15,7 +14,7 @@ from apps.content.services.validate_recitation_tracks_upload_service import (
     ValidateRecitationTracksUploadService,
 )
 from apps.core.ninja_utils.auth import ninja_jwt_auth
-from apps.core.ninja_utils.errors import ItqanError, NinjaErrorResponse
+from apps.core.ninja_utils.errors import NinjaErrorResponse
 from apps.core.ninja_utils.request import Request
 from apps.core.ninja_utils.router import ItqanRouter
 from apps.core.ninja_utils.tags import NinjaTag
@@ -240,34 +239,3 @@ def abort_upload(request: Request, data: UploadAbortIn):
     result = service.abort_upload(key=data.key, upload_id=data.upload_id)
 
     return UploadAbortOut(key=data.key, upload_id=data.upload_id, aborted=result["aborted"])
-
-
-class DeleteTracksIn(Schema):
-    track_ids: list[int]
-
-
-@router.delete(
-    "recitation-tracks/",
-    response={
-        204: None,
-        401: NinjaErrorResponse[Literal["authentication_error"], Literal[None]],
-        403: NinjaErrorResponse[Literal["permission_denied"], Literal[None]],
-    },
-    auth=ninja_jwt_auth,
-)
-def delete_tracks(request: Request, data: DeleteTracksIn):
-    if not request.user.is_staff:
-        raise PermissionDenied("Staff only")
-
-    repo = RecitationTrackRepository()
-    tracks = repo.get_recitation_tracks_by_ids(data.track_ids)
-
-    if not data.track_ids or len(tracks) != len(data.track_ids):
-        raise ItqanError(
-            error_name="track_not_found",
-            message="Some track IDs are invalid or do not exist.",
-            status_code=400,
-        )
-
-    repo.delete_recitation_tracks(tracks)
-    return 204, None
